Unbreak all the views; refactor and fix warnings; allow the popover to overflow the schedule grid
